+++ src/rules/index.ts
@@ -1,2 +1,2 @@
 export {Rule, RuleOperator, RuleVariables} from './types';
-export {checkRules} from './check';
+export {checkRules} from './check-rules';
